Abstract TSO communication protocol on the client-side in an interface
No functional changes added

Also:
- Protocol exception classes extracted from TSOClient
- Add javadoc to interface extracted from TSOClient
- Merge TSOClient and TSOClientImpl in a single class
