+++ transaction-client/src/main/java/com/yahoo/omid/tsoclient/ClosingException.java
@@ -1,0 +1,26 @@
+/**
+ * Copyright 2011-2016 Yahoo Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *         http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package com.yahoo.omid.tsoclient;
+
+/**
+ * Thrown when an error is produced when performing the actions required
+ * to close the communication with the TSO server
+ */
+public class ClosingException extends Exception {
+
+    private static final long serialVersionUID = -5681694952053689884L;
+
+}
