+++ transaction-client/src/main/java/com/yahoo/omid/tsoclient/ConnectionException.java
@@ -1,0 +1,28 @@
+/**
+ * Copyright 2011-2016 Yahoo Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *         http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package com.yahoo.omid.tsoclient;
+
+import java.io.IOException;
+
+/**
+ * Thrown when there are problems with the comm channel with the TSO server
+ * (e.g. when it is closed or disconnected)
+ */
+public class ConnectionException extends IOException {
+
+    private static final long serialVersionUID = -8489539195700267443L;
+
+}
