+++ transaction-client/src/main/java/com/yahoo/omid/tsoclient/TSOProtocol.java
@@ -1,0 +1,54 @@
+/**
+ * Copyright 2011-2016 Yahoo Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *         http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package com.yahoo.omid.tsoclient;
+
+import java.util.Set;
+
+/**
+ * Defines the protocol used on the client side to abstract communication to the TSO server
+ */
+public interface TSOProtocol {
+
+    /**
+     * Returns a new timestamp assigned by on the server-side
+     * @return the newly assigned timestamp as a future. If an error was detected, the future will contain a
+     * corresponding protocol exception
+     * @see TimestampOracle
+     * @see TSOServer
+     */
+    TSOFuture<Long> getNewStartTimestamp();
+
+    /**
+     * Returns the result of the conflict detection made on the server-side for the specified transaction
+     * @param transactionId
+     *          the transaction to check for conflicts
+     * @param writeSet
+     *          the writeSet of the transaction, which includes all the modified cells
+     * @return the commit timestamp as a future if the transaction was committed. If the transaction was aborted due
+     * to conflicts with a concurrent transaction, the future will include an AbortException. If an error was detected,
+     * the future will contain a corresponding protocol exception
+     * @see TimestampOracle
+     * @see TSOServer
+     */
+    TSOFuture<Long> commit(long transactionId, Set<? extends CellId> writeSet);
+
+    /**
+     * Closes the communication with the TSO server
+     * @return nothing. If an error was detected, the future will contain a corresponding protocol exception
+     */
+    TSOFuture<Void> close();
+
+}
