+++ tso-server/src/test/java/com/yahoo/omid/tsoclient/TSOClientAccessor.java
@@ -1,13 +1,12 @@
 package com.yahoo.omid.tsoclient;
 
-import com.yahoo.omid.tsoclient.TSOClientImpl.ConnectedState;
+import com.yahoo.omid.tsoclient.TSOClient.ConnectedState;
 import com.yahoo.statemachine.StateMachine.FsmImpl;
 
 public class TSOClientAccessor {
 
     public static void closeChannel(TSOClient tsoClient) throws InterruptedException {
-        TSOClientImpl clientimpl = (TSOClientImpl) tsoClient;
-        FsmImpl fsm = (FsmImpl) clientimpl.fsm;
+        FsmImpl fsm = (FsmImpl) tsoClient.fsm;
         ConnectedState connectedState = (ConnectedState) fsm.getState();
         connectedState.channel.close().await();
     }
